Crate configuration

Modules

shared

Structs

GlobalSettings

Global settings applied to an entire network.

GlobalSettingsBuilder

A global settings builder, used to build GlobalSettings declaratively with fields validation.

HrmpChannelConfig

HRMP channel configuration, with fine-grained configuration options.

HrmpChannelConfigBuilder

HRMP channel configuration builder, used to build an HrmpChannelConfig declaratively with fields validation.

NetworkConfig

A network configuration, composed of a relaychain, parachains and HRMP channels.

NetworkConfigBuilder

A network configuration builder, used to build a NetworkConfig declaratively with fields validation.

ParachainConfig

A parachain configuration, composed of collators and fine-grained configuration options.

ParachainConfigBuilder

A parachain configuration builder, used to build a ParachainConfig declaratively with fields validation.

RelaychainConfig

A relay chain configuration, composed of nodes and fine-grained configuration options.

RelaychainConfigBuilder

A relay chain configuration builder, used to build a RelaychainConfig declaratively with fields validation.